Update on Dec, 16.

+++ src/com/manu/dynasty/template/YunBiaoSafe.java
@@ -2,7 +2,7 @@ package com.manu.dynasty.template;
 
 public class YunBiaoSafe {
 	public int areaID;
-	public int saveAreaX;
-	public int saveAreaZ;
-	public int saveArear;
+	public float saveAreaX;
+	public float saveAreaZ;
+	public float saveArear;
 }
